shipprimus_app/requirements.txt
flask
flask-session
requests
requests-oauthlib
python-dotenv
reportlab
PyPDF2
